extract ld encoder family
